backend: move functions code
