+++ apps/backend-functions/src/delete.ts
