+++ apps/backend-functions/src/generateHash.ts
