+++ apps/backend-functions/tsconfig.json
@@ -16,6 +16,7 @@
       "./node_modules/@types"
     ]
   },
+  "compileOnSave": true,
   "include": [
     "src"
   ]
