--- functions/.gitignore
@@ -1,10 +1,0 @@
-## Compiled JavaScript files
-**/*.js
-**/*.js.map
-lib/contracts/*
-
-# Typescript v1 declaration files
-typings/
-
-node_modules/
-coverage/
