--- functions/package.json
@@ -1,59 +1,0 @@
-{
-  "name": "functions",
-  "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "test": "jest --coverage",
-    "test:watch": "jest --watchAll --coverage",
-    "build": "tsc",
-    "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "@google-cloud/storage": "^2.4.2",
-    "@types/tmp": "0.0.34",
-    "ethers": "^4.0.27",
-    "firebase": "^6.0.2",
-    "firebase-admin": "^7.0.0",
-    "firebase-functions": "^2.1.0",
-    "line-reader": "^0.4.0",
-    "lodash": "^4.17.11",
-    "request": "^2.88.0",
-    "tmp": "0.0.33"
-  },
-  "devDependencies": {
-    "@types/jest": "^23.3.14",
-    "@types/line-reader": "0.0.28",
-    "@types/lodash": "^4.14.123",
-    "jest": "^23.6.0",
-    "ts-jest": "^23.10.5",
-    "tslint": "^5.12.0",
-    "typescript": "^3.2.2"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testRegex": "((\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    "testPathIgnorePatterns": [
-      "<rootDir>/node_modules",
-      "<rootDir>/lib",
-      "<rootDir>/coverage"
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ]
-  },
-  "private": true
-}
