--- functions/test.json
@@ -1,11 +1,0 @@
-{
-    "director": "",
-    "title": "",
-    "logline": "",
-    "principalInformations": {
-        "producers": "",
-        "keywords": [
-            "", "",""
-        ]
-    }
-}
