--- functions/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "outDir": "lib",
-    "sourceMap": true,
-    "strict": true,
-    "target": "es2015",
-    "typeRoots" : ["./node_modules/@types"]
-  },
-  "compileOnSave": true,
-  "include": [
-    "src"
-  ]
-}
